[fix] GetUniqueId中rmtBufferVec.size()显式转换为u32，保持与BufferVecPack序列化一致

+++ src/legacy/unified_platform/resource/transport/ub_mem_transport.cc
@@ -786,7 +786,7 @@ std::vector<char> UbMemTransport::GetUniqueId()
     binaryStream << type;
     binaryStream << notifyNum;
     binaryStream << bufferNum;
-    binaryStream << rmtBufferVec.size();
+    binaryStream << static_cast<u32>(rmtBufferVec.size());
     binaryStream << connNum;
 
     // [header...][notifyUniqueId...][rmtNotifyUniqueId...][rmtBufferUniqueIds...]
